+++ container-search/src/main/java/ai/vespa/llm/clients/ConfigurableLanguageModel.java
@@ -1,8 +1,8 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package ai.vespa.llm.clients;
 
-import ai.vespa.llm.InferenceParameters;
-import ai.vespa.llm.LanguageModel;
+import ai.vespa.languagemodels.InferenceParameters;
+import ai.vespa.languagemodels.LanguageModel;
 import ai.vespa.llm.LlmClientConfig;
 import com.yahoo.api.annotations.Beta;
 import com.yahoo.component.annotation.Inject;
@@ -19,7 +19,7 @@ import java.util.logging.Logger;
 @Beta
 public abstract class ConfigurableLanguageModel implements LanguageModel {
 
-    private static Logger log = Logger.getLogger(ai.vespa.llm.clients.ConfigurableLanguageModel.class.getName());
+    private static Logger log = Logger.getLogger(ConfigurableLanguageModel.class.getName());
 
     private final String apiKey;
     private final String endpoint;
@@ -68,7 +68,9 @@ public abstract class ConfigurableLanguageModel implements LanguageModel {
     }
 
     protected void setEndpoint(InferenceParameters params) {
-        params.setEndpoint(endpoint);
+        if (endpoint != null && ! endpoint.isEmpty()) {
+            params.setEndpoint(endpoint);
+        }
     }
 
 }
