+++ container-search/src/main/java/ai/vespa/llm/search/RAGSearcher.java
@@ -1,10 +1,10 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package ai.vespa.llm.search;
 
-import ai.vespa.llm.LanguageModel;
+import ai.vespa.languagemodels.LanguageModel;
 import ai.vespa.llm.LlmSearcherConfig;
-import ai.vespa.llm.completion.Prompt;
-import ai.vespa.llm.completion.StringPrompt;
+import ai.vespa.languagemodels.completion.Prompt;
+import ai.vespa.languagemodels.completion.StringPrompt;
 import com.yahoo.api.annotations.Beta;
 import com.yahoo.component.annotation.Inject;
 import com.yahoo.component.provider.ComponentRegistry;
