+++ container-search/src/test/java/ai/vespa/llm/clients/ConfigurableLanguageModelTest.java
@@ -1,11 +1,11 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package ai.vespa.llm.clients;
 
-import ai.vespa.llm.InferenceParameters;
+import ai.vespa.languagemodels.InferenceParameters;
 import ai.vespa.llm.LlmClientConfig;
-import ai.vespa.llm.completion.Completion;
-import ai.vespa.llm.completion.Prompt;
-import ai.vespa.llm.completion.StringPrompt;
+import ai.vespa.languagemodels.completion.Completion;
+import ai.vespa.languagemodels.completion.Prompt;
+import ai.vespa.languagemodels.completion.StringPrompt;
 import com.yahoo.container.di.componentgraph.Provider;
 import com.yahoo.container.jdisc.SecretStoreProvider;
 import com.yahoo.container.jdisc.secretstore.SecretStore;
