+++ vespajlib/src/main/java/ai/vespa/languagemodels/InferenceParameters.java
@@ -1,5 +1,5 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.llm;
+package ai.vespa.languagemodels;
 
 import com.yahoo.api.annotations.Beta;
