+++ vespajlib/src/main/java/ai/vespa/languagemodels/LanguageModel.java
@@ -1,8 +1,8 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.llm;
+package ai.vespa.languagemodels;
 
-import ai.vespa.llm.completion.Completion;
-import ai.vespa.llm.completion.Prompt;
+import ai.vespa.languagemodels.completion.Completion;
+import ai.vespa.languagemodels.completion.Prompt;
 import com.yahoo.api.annotations.Beta;
 
 import java.util.List;
