+++ vespajlib/src/main/java/ai/vespa/languagemodels/client/openai/package-info.java
@@ -1,11 +1,11 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 @ExportPackage
 @PublicApi
-package ai.vespa.llm.completion;
+package ai.vespa.languagemodels.client.openai;
 
 import com.yahoo.api.annotations.PublicApi;
 import com.yahoo.osgi.annotation.ExportPackage;
 
 /**
- * Classes for generating text completions with language models.
+ * Client to OpenAi's large language models.
  */
