+++ vespajlib/src/main/java/ai/vespa/languagemodels/package-info.java
@@ -1,11 +1,11 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 @ExportPackage
 @PublicApi
-package ai.vespa.llm.client.openai;
+package ai.vespa.languagemodels;
 
 import com.yahoo.api.annotations.PublicApi;
 import com.yahoo.osgi.annotation.ExportPackage;
 
 /**
- * Client to OpenAi's large language models.
+ * API for working with large language models.
  */
