+++ vespajlib/src/main/java/ai/vespa/languagemodels/test/package-info.java
@@ -1,7 +1,7 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 @ExportPackage
 @PublicApi
-package ai.vespa.llm.test;
+package ai.vespa.languagemodels.test;
 
 /**
  * Tools for writing tests when working with large language models.
